refactor: relocate transitions and fonts imports
Moved `transitions` from `styles` to `utils` and `fonts` from `styles` to `lib/nextjs`. Updated all related imports for consistency. This improves overall code structure and aligns with established modular conventions.
