+++ apps/nextjs/src/components/layout/navigation/footer/help.tsx
@@ -5,7 +5,7 @@ import { Separator } from '@/components/layout/navigation/separator';
 import { Icon } from '@/components/ui/icon';
 import { List } from '@/components/ui/list';
 import { Text } from '@/components/ui/text';
-import { transitions } from '@/styles/transitions';
+import { transitions } from '@/utils/transitions';
 import { PADDING_X } from '../navigation';
 
 export const Help: React.FC = () => {
