+++ apps/nextjs/src/features/inbox/components/inbox-list-header/inbox-list-header.tsx
@@ -1,7 +1,7 @@
 import { memo, useEffect, useMemo, useState } from 'react';
 import { useInView } from 'react-intersection-observer';
 import { Flex, type FlexProps } from '@/components/ui/flex';
-import { transitions } from '@/styles/transitions';
+import { transitions } from '@/utils/transitions';
 
 type Props = FlexProps;
