+++ apps/nextjs/src/features/task-detail/components/attachment-box/component.tsx
@@ -2,8 +2,8 @@ import { Flex, type FlexProps } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
 import { Link } from '@/components/ui/link';
 import { Text } from '@/components/ui/text';
-import { transitions } from '@/styles/transitions';
 import type { IconType } from '@/utils/icons';
+import { transitions } from '@/utils/transitions';
 import { type Sizes, sizes } from './sizes';
 
 type Props = FlexProps & {
