+++ apps/nextjs/src/features/task-detail/components/attachment-uploading-box/attachment-uploading-box.tsx
@@ -2,7 +2,7 @@ import { Flex, type FlexProps } from '@/components/ui/flex';
 import { Icon } from '@/components/ui/icon';
 import { Progress } from '@/components/ui/progress';
 import { Text } from '@/components/ui/text';
-import { transitions } from '@/styles/transitions';
+import { transitions } from '@/utils/transitions';
 
 type Props = FlexProps & {
   size: Sizes;
