+++ apps/nextjs/src/features/task/components/tasks-board/tasks-board-content/tasks-board-content.tsx
@@ -5,8 +5,8 @@ import { useTaskDetailDrawer } from '@/features/task-detail/components/task-deta
 import { useMainStyle } from '@/hooks/styles/use-main-style';
 import { usePrevious } from '@/hooks/use-previous';
 import { useBreakpointValue } from '@/lib/chakra';
-import { transitions } from '@/styles/transitions';
 import { isHTMLElement } from '@/utils/is-html-element';
+import { transitions } from '@/utils/transitions';
 
 type Props = FlexProps;
