+++ apps/nextjs/src/features/task/components/tasks-calendar/tasks-calendar-list-item/list-item/container/container.tsx
@@ -6,7 +6,7 @@ import {
   useProjectBaseColorText,
 } from '@/store/entities/project-base-color';
 import { useTask } from '@/store/entities/task';
-import { transitions } from '@/styles/transitions';
+import { transitions } from '@/utils/transitions';
 
 type Props = {
   taskId: string;
