+++ apps/nextjs/src/features/task/components/tasks-files/tasks-files-content/tasks-files-content.tsx
@@ -1,7 +1,7 @@
 import { memo } from 'react';
 import { Flex, type FlexProps } from '@/components/ui/flex';
 import { useMainStyle } from '@/hooks/styles/use-main-style';
-import { transitions } from '@/styles/transitions';
+import { transitions } from '@/utils/transitions';
 
 type Props = FlexProps;
