+++ apps/nextjs/src/features/task/components/tasks-files/tasks-files-list-item/tasks-files-list-item.tsx
@@ -16,7 +16,7 @@ import {
   useTaskFile,
   useTaskFileIdsByTaskId,
 } from '@/store/entities/task-file';
-import { transitions } from '@/styles/transitions';
+import { transitions } from '@/utils/transitions';
 
 type Props = {
   taskFileId: string;
